+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Added
     - Custom password argument in ``createdevusers`` management command (#1393)
     - ``PluginObjectLink`` data class in plugins (#1343)
     - ``PluginSearchResult`` data class in plugins (#1399)
+    - Target user ``sodar_uuid`` updating in remote sync (#1316, #1317)
+    - Update local user data in remote sync (#1407)
 
 Changed
 -------